i2n: Restore the local aexpect module
This partially reverts the changes of commit 4266971159 which
removes the aexpect module in favor of external repo for it.
However, this is not done in virt-test (where there is another
copy of the aexpect code) and the external rpm is not immediately
available on all our workstations. Thus, let's wait for it to
mature a bit and take the simple step of of providing the old
module for now.